Report error when cquery fails
When the plugin is invoked without providing a list of targets to
build, it tries to infer the targets using a cquery like so:

bazel build $(bazel cquery 'rdeps(...)')

In certain cases, the cquery can fail with a non-zero exit code
and return an empty result. For example: if there's no target
corresponding to the file from where the plugin was invoked.

In such cases, the outer call to build ignores the non-zero exit
code and proceeds to build an empty target list and succeeds.

This is a problem because the plugin makes it appear that the build
has succeeded when nothing was ever built.

There are some nice ways of solving this problem:
- set -euo pipefail -- to catch the subcommand failure and exit
- Separate the cquery and build commands, check the cquery exit code
  separately and report errors

Both these solutions are not practical for this plugin because:
- I don't know how to do "set -euo pipefail" with th compiler plugin
- Separating cquery and build increases the code complexity because
  we now need to decide if both will be done async or not. If yes, then
  there will be multiple calls to the async plugin. If no, then the
  cquery stage will block the editor

This commit uses a third approach where we return a nonsensical target
(CQUERY_FAILED) name when cquery exits with a non-zero status. This
causes the outer build command to fail too albeit with a somewhat
confusing error message.